Animations + Sprites
There WILL BE merge conflicts; I had to recreate the code to move the player. Dante said this was okay to do, and I have the original code backed up on a notepad if we want to go back.

Not all sprites are added; I will wait to add any more until this project is in a more completed state.